+++ src/torchmetrics/text/__init__.py
@@ -14,6 +14,7 @@
 from torchmetrics.text.bleu import BLEUScore
 from torchmetrics.text.cer import CharErrorRate
 from torchmetrics.text.chrf import CHRFScore
+from torchmetrics.text.edit import EditDistance
 from torchmetrics.text.eed import ExtendedEditDistance
 from torchmetrics.text.mer import MatchErrorRate
 from torchmetrics.text.perplexity import Perplexity
@@ -24,9 +25,9 @@ from torchmetrics.text.ter import TranslationEditRate
 from torchmetrics.text.wer import WordErrorRate
 from torchmetrics.text.wil import WordInfoLost
 from torchmetrics.text.wip import WordInfoPreserved
-from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
+from torchmetrics.utilities.imports import _TRANSFORMERS_GREATER_EQUAL_4_4
 
-if _TRANSFORMERS_AVAILABLE:
+if _TRANSFORMERS_GREATER_EQUAL_4_4:
     from torchmetrics.text.bert import BERTScore
     from torchmetrics.text.infolm import InfoLM
 
@@ -35,6 +36,7 @@ __all__ = [
     "BLEUScore",
     "CharErrorRate",
     "CHRFScore",
+    "EditDistance",
     "ExtendedEditDistance",
     "MatchErrorRate",
     "Perplexity",